refactor: moved commands into their own seperate modules
