+++ src/Session/Session.ts
@@ -1,25 +1,24 @@
 import uuidv1 from 'uuid/v1';
-import validator from 'validator';
 import os from 'os';
 import { Mutex } from 'async-mutex';
-import { VM } from 'vm2';
 import { JSDOM } from 'jsdom';
 import has from 'has';
 
 import { WebElement } from '../WebElement/WebElement';
-import { COMMANDS, ELEMENT } from '../constants/constants';
+import { COMMANDS } from '../constants/constants';
 import { Browser } from '../Browser/Browser';
 import { Pluma } from '../Types/types';
+import { sendKeysToElement } from './SendKeysToElement';
+import { navigateTo } from './NavigateTo';
+import { addElementToKnownElements } from './AddToKnownElements';
 import * as utils from '../utils/utils';
 
 // custom
-import { addFileList } from '../jsdom_extensions/addFileList';
 
 // DOM specific
 const {
   /** A window event, imported from jsdom */
 
-  Event,
   /** jsdom implementation of the HTMLElement object */
 
   HTMLElement,
@@ -30,13 +29,11 @@ import {
   InvalidArgument,
   SessionNotCreated,
   NoSuchElement,
-  ElementNotInteractable,
   NoSuchWindow,
-  JavaScriptError,
-  ScriptTimeout,
 } from '../Error/errors';
 
 import { CapabilityValidator } from '../CapabilityValidator/CapabilityValidator';
+import { executeScript } from './ExecuteScript';
 
 /**
  * Represents the connection between a local end and a specific remote end. In this case, jsdom.
@@ -90,7 +87,7 @@ class Session {
         await this.browser.close();
         break;
       case COMMANDS.NAVIGATE_TO:
-        await this.navigateTo(parameters.url);
+        await navigateTo(parameters.url, this.browser, this.timeouts);
         break;
       case COMMANDS.GET_CURRENT_URL:
         response = this.browser.getUrl();
@@ -174,14 +171,20 @@ class Session {
         break;
       case COMMANDS.EXECUTE_SCRIPT:
         if (!this.browser.dom.window) throw new NoSuchWindow();
-        const value: unknown = await this.executeScript(
+        const value: unknown = await executeScript(
           parameters.script,
           parameters.args,
+          this.browser,
+          this.timeouts,
         );
         response = { value };
         break;
       case COMMANDS.ELEMENT_SEND_KEYS:
-        await this.sendKeysToElement(parameters.text, urlVariables.elementId);
+        await sendKeysToElement(
+          parameters.text,
+          urlVariables.elementId,
+          this.browser,
+        );
         break;
       case COMMANDS.ELEMENT_CLICK:
         if (!this.browser.dom.window) throw new NoSuchWindow();
@@ -213,8 +216,9 @@ class Session {
         break;
       case COMMANDS.GET_ACTIVE_ELEMENT:
         if (!this.browser.dom.window) throw new NoSuchWindow();
-        response = this.addElementToKnownElements(
+        response = addElementToKnownElements(
           this.browser.getActiveElement(),
+          this.browser,
         );
         break;
       case COMMANDS.SWITCH_TO_FRAME:
@@ -237,104 +241,6 @@ class Session {
     }
     return response || { value: null };
   }
-
-  /**
-   * Accepts a string and an elementId @type {string}
-   * Tries to locate the element with the user provided Id and insert the specified string of text
-   * sets a user defined value on a given HTML element
-   * TODO: this method needs to be updated to incorporate the action Object
-   */
-  sendKeysToElement(text: string, elementId: string): Promise<void> {
-    return new Promise(async (resolve, reject) => {
-      const webElement = this.browser.getKnownElement(elementId);
-      const element: HTMLElement = webElement.element;
-      let files = [];
-
-      if (text === undefined) reject(new InvalidArgument());
-
-      if (
-        !webElement.isInteractable() &&
-        element.getAttribute('contenteditable') !== 'true'
-      ) {
-        reject(new ElementNotInteractable()); // TODO: create new error class
-      }
-
-      if (this.browser.getActiveElement() !== element) element.focus();
-
-      if (element.tagName.toLowerCase() === 'input') {
-        if (typeof text === 'string') reject(new InvalidArgument());
-        // file input
-        if (element.getAttribute('type') === 'file') {
-          files = text.split('\n');
-          if (files.length === 0) throw new InvalidArgument();
-          if (!element.hasAttribute('multiple') && files.length !== 1)
-            throw new InvalidArgument();
-
-          await Promise.all(
-            files.map(file => utils.fileSystem.pathExists(file)),
-          );
-
-          addFileList(element, files);
-          element.dispatchEvent(new Event('input'));
-          element.dispatchEvent(new Event('change'));
-        } else if (
-          element.getAttribute('type') === 'text' ||
-          element.getAttribute('type') === 'email'
-        ) {
-          (element as HTMLInputElement).value += text;
-          element.dispatchEvent(new Event('input'));
-          element.dispatchEvent(new Event('change'));
-        } else if (element.getAttribute('type') === 'color') {
-          if (!validator.isHexColor(text)) throw new InvalidArgument();
-          (element as HTMLInputElement).value = text;
-        } else {
-          if (!has(element, 'value') || element.getAttribute('readonly'))
-            throw new Error('element not interactable'); // TODO: create error class
-          // TODO: add check to see if element is mutable, reject with element not interactable
-          (element as HTMLInputElement).value = text;
-        }
-        element.dispatchEvent(new Event('input'));
-        element.dispatchEvent(new Event('change'));
-        resolve(null);
-      } else {
-        // TODO: text needs to be encoded before it is inserted into the element
-        // innerHTML, especially important since js code can be inserted in here and executed
-        element.innerHTML += text;
-        resolve(null);
-      }
-    });
-  }
-
-  /**
-   * navigates to a specified url
-   * sets timers according to session config
-   */
-  async navigateTo(url: string): Promise<void> {
-    let pathType: string;
-
-    try {
-      if (validator.isURL(url)) pathType = 'url';
-      else if (await utils.fileSystem.pathExists(url)) pathType = 'file';
-      else throw new InvalidArgument();
-    } catch (e) {
-      throw new InvalidArgument();
-    }
-
-    // pageload timer
-    let timer;
-    const startTimer = (): void => {
-      timer = setTimeout(() => {
-        throw new Error('timeout'); // TODO: create timeout error class
-      }, this.timeouts.pageLoad);
-    };
-
-    if (this.browser.getUrl() !== url) {
-      startTimer();
-      await this.browser.navigate(url, pathType);
-      clearTimeout(timer);
-    }
-  }
-
   /**
    * sets and validates the [[timeouts]] object
    * */
@@ -661,92 +567,6 @@ class Session {
       this.browser.knownElements.push(foundElement);
     });
     return result;
-  }
-
-  /**
-   * handles errors resulting from failing to execute synchronous scripts
-   */
-  private handleSyncScriptError({
-    message,
-    code,
-  }: NodeJS.ErrnoException): never {
-    if (code === 'ERR_SCRIPT_EXECUTION_TIMEOUT') {
-      throw new ScriptTimeout();
-    } else {
-      throw new JavaScriptError(message);
-    }
-  }
-
-  /**
-   * Adds an HTMLElement to the array of known elements
-   * @param {HTMLElement} element
-   * @returns {Object} the JSON representation of the WebElement
-   */
-  private addElementToKnownElements(
-    element: HTMLElement,
-  ): Pluma.SerializedWebElement {
-    const webElement = new WebElement(element);
-    this.browser.knownElements.push(webElement);
-    return webElement.serialize();
-  }
-
-  /**
-   * executes a user defined script within the context of the dom on a given set of user defined arguments
-   */
-  public executeScript(script: string, args: unknown[]): unknown {
-    const argumentList = args.map(arg => {
-      if (arg[ELEMENT] == null) {
-        return arg;
-      } else {
-        const { element } = this.browser.getKnownElement(arg[ELEMENT]);
-        return element;
-      }
-    });
-
-    const window = this.browser.getCurrentBrowsingContextWindow();
-
-    const func = window
-      .eval(`(function() {${script}})`)
-      .bind(null, ...argumentList);
-
-    const vm = new VM({
-      timeout: this.timeouts.script,
-      sandbox: {
-        func,
-      },
-    });
-
-    let vmReturnValue;
-
-    try {
-      vmReturnValue = vm.run('func();');
-    } catch (error) {
-      this.handleSyncScriptError(error);
-    }
-
-    // TODO: incorporate @types/jsdom to resolve typescript instanceof errors
-    // eslint-disable-next-line
-    const { NodeList, HTMLCollection, HTMLElement } = window as any;
-
-    if (
-      vmReturnValue instanceof NodeList ||
-      vmReturnValue instanceof HTMLCollection
-    ) {
-      vmReturnValue = Array.from(vmReturnValue);
-    }
-
-    if (Array.isArray(vmReturnValue)) {
-      return vmReturnValue.map(value =>
-        value instanceof HTMLElement
-          ? this.addElementToKnownElements(value)
-          : value,
-      );
-    } else if (vmReturnValue instanceof HTMLElement) {
-      return this.addElementToKnownElements(vmReturnValue);
-    }
-
-    // client will expect undefined return values to be null
-    return typeof vmReturnValue === 'undefined' ? null : vmReturnValue;
   }
 }
